OrderNum字段由Integer调整为Long, 增加EnabledEntity接口

+++ guerlab-cloud-commons/src/test/java/net/guerlab/cloud/commons/util/TreeUtilsTests.java
@@ -91,9 +91,9 @@ class TreeUtilsTests {
 
 		private Long parentId;
 
-		private Integer orderNum;
+		private Long orderNum;
 
-		TestEntity1(Long id, Long parentId, int orderNum) {
+		TestEntity1(Long id, Long parentId, long orderNum) {
 			this.id = id;
 			this.parentId = parentId;
 			this.orderNum = orderNum;
